Add Rio SEO storefinder
Updated Hallmark spider to use new Rio SEO storefinder
Added See's Candies with new Rio SEO storefinder
